requirements.txt
Babel==1.3
defusedxml==0.4.1
Django==1.8.4
django-admin-bootstrapped==2.5.6
django-allauth==0.23.0
django-appconf==1.0.1
django-braces==1.8.1
django-compressor==1.5
django-core==1.4.1
django-crispy-forms==1.5.2
django-extensions==1.5.7
django-extra-views==0.6.4
django-tables2==1.0.4
django-treebeard==3.0
django-widget-tweaks==1.3
djangorestframework==3.2.4
factory-boy==2.4.1
google-api-python-client==1.4.2
httplib2==0.9.1
mock==1.0.1
oauth2client==1.5.1
oauthlib==1.0.3
phonenumbers==6.3.0
Pillow==2.7.0
purl==1.1
pyasn1==0.1.8
pyasn1-modules==0.0.7
pycountry==1.17
python3-openid==3.0.7
pytz==2015.6
requests==2.7.0
requests-oauthlib==0.5.0
rsa==3.2
simplejson==3.8.0
six==1.9.0
sorl-thumbnail==12.2
Unidecode==0.4.18
uritemplate==0.6
wheel==0.24.0
